+++ lib/index.d.ts
@@ -1,8 +1,0 @@
-import * as Async from "./async";
-export { Async };
-import * as Sync from "./sync";
-export { Sync };
-import * as Pipeline from "./pipeline";
-export { Pipeline };
-export { RecPartial } from "./shared";
-export { FactoryFunc, Generator, Derived, Factory, Builder, val, each, makeFactory } from "./sync";