+++ spec/async.spec.ts
@@ -3,14 +3,14 @@ import * as Sync from "../src/sync";
 import { makeFactory } from "../src/async";
 
 interface ParentType {
-  name: string;
+  name: string | null;
   birthday: Date;
   children: ChildType[];
   spouse: ParentType | null;
 }
 
 interface ChildType {
-  name: string;
+  name: string | null;
   grade: number;
 }
 
@@ -34,6 +34,11 @@ describe("async factories build stuff", () => {
     const jimmy = await childFactory.build();
     expect(jimmy.name).toEqual("Kid");
     expect(jimmy.grade).toEqual(1);
+  });
+  it("makes an object with default field explicitly set to null", async () => {
+    const anon = await childFactory.build({ name: null });
+    expect(anon.name).toBeNull();
+    expect(anon.grade).toEqual(1);
   });
   it("can make use of sequence #", async () => {
     const susan = await parentFactory.build({ name: "Susan" });
@@ -264,5 +269,21 @@ describe("async factories build stuff", () => {
     }, { startingSequenceNumber: 3 });
     const a = await factoryA.build();
     expect(a.foo).toEqual(4);
-  })
+  });
+  it("clones deeply nested values", async () => {
+    interface TypeA {
+      bar: {
+        baz: string;
+      };
+    }
+    const factoryA = Async.makeFactory<TypeA>({
+      bar: {
+        baz: "should-be-immutable"
+      }
+    });
+    const a = await factoryA.build();
+    const b = await factoryA.build();
+    a.bar.baz = "is-not-immutable";
+    expect(b.bar.baz).toEqual("should-be-immutable");
+  });
 });